move utils to test_common
